Revert "feat: custom user agent Storage changes for UI handoff (#11627)"
This reverts commit b0231af9d7fe631ef9e0e669df7a20802e3a21b3.
