+++ packages/core/src/Platform/types.ts
@@ -127,7 +127,6 @@ export enum StorageAction {
 	Copy = '4',
 	Remove = '5',
 	GetProperties = '6',
-	Cancel = '7',
 }
 
 type ActionMap = {
