+++ packages/storage/src/Storage.ts
@@ -1,7 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { Amplify, ConsoleLogger as Logger } from '@aws-amplify/core';
+import {
+	Amplify,
+	ConsoleLogger as Logger,
+	parseAWSExports,
+} from '@aws-amplify/core';
 import { AWSS3Provider } from './providers';
 import {
 	StorageCopySource,
@@ -23,7 +27,8 @@ import {
 	StorageGetPropertiesOutput,
 } from './types';
 import { PutObjectInput } from './AwsClients/S3';
-import { InternalStorageClass } from './internals/InternalStorage';
+import { isCancelError } from './AwsClients/S3/utils';
+import { AWSS3UploadTask } from './providers/AWSS3UploadTask';
 
 const logger = new Logger('StorageClass');
 const loggerStorageInstance = new Logger('Storage'); // Logging relating to Storage instance management
@@ -32,14 +37,173 @@ const DEFAULT_PROVIDER = 'AWSS3';
 /**
  * Provide storage methods to use AWS S3
  */
-export class Storage extends InternalStorageClass {
+export class Storage {
+	/**
+	 * @private
+	 */
+	private _config;
+	private _pluggables: StorageProvider[];
+
+	/**
+	 * Similar to the API module. This weak map allows users to cancel their in-flight request made using the Storage
+	 * module. For every get or put request, a unique AbortConttroller will be generated and injected to it's underlying
+	 * Xhr HTTP handler. This map maintains a mapping of Request to AbortController. When .cancel is invoked, it will
+	 * attempt to retrieve it's corresponding abortController and cancel the in-flight request.
+	 */
+	private _abortControllerMap: WeakMap<Promise<any>, AbortController>;
+
 	/**
 	 * @public
 	 */
 	public vault: Storage;
 
+	/**
+	 * Initialize Storage
+	 * @param {Object} config - Configuration object for storage
+	 */
+	constructor() {
+		this._config = {};
+		this._pluggables = [];
+		this._abortControllerMap = new WeakMap<Promise<any>, AbortController>();
+		logger.debug('Storage Options', this._config);
+
+		this.get = this.get.bind(this);
+		this.put = this.put.bind(this);
+		this.remove = this.remove.bind(this);
+		this.list = this.list.bind(this);
+	}
+
 	public getModuleName() {
 		return 'Storage';
+	}
+
+	/**
+	 * add plugin into Storage category
+	 * @param {Object} pluggable - an instance of the plugin
+	 */
+	public addPluggable(pluggable: StorageProvider) {
+		if (pluggable && pluggable.getCategory() === 'Storage') {
+			this._pluggables.push(pluggable);
+			let config = {};
+
+			config = pluggable.configure(this._config[pluggable.getProviderName()]);
+
+			return config;
+		}
+	}
+
+	/**
+	 * Get the plugin object
+	 * @param providerName - the name of the plugin
+	 */
+	public getPluggable(providerName: string) {
+		const pluggable = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === providerName
+		);
+		if (pluggable === undefined) {
+			logger.debug('No plugin found with providerName', providerName);
+			return null;
+		} else return pluggable;
+	}
+
+	/**
+	 * Remove the plugin object
+	 * @param providerName - the name of the plugin
+	 */
+	public removePluggable(providerName: string) {
+		this._pluggables = this._pluggables.filter(
+			pluggable => pluggable.getProviderName() !== providerName
+		);
+		return;
+	}
+
+	/**
+	 * Configure Storage
+	 * @param {Object} config - Configuration object for storage
+	 * @return {Object} - Current configuration
+	 */
+	configure(config?) {
+		logger.debug('configure Storage');
+		if (!config) return this._config;
+
+		const amplifyConfig = parseAWSExports(config);
+
+		const storageConfig = amplifyConfig.Storage ?? {};
+
+		const defaultProviderConfigKeys = [
+			'bucket',
+			'region',
+			'level',
+			'track',
+			'customPrefix',
+			'ContentMD5',
+			'serverSideEncryption',
+			'SSECustomerAlgorithm',
+			'SSECustomerKey',
+			// TODO(AllanZhengYP): remove in V6.
+			'SSECustomerKeyMD5',
+			'SSEKMSKeyId',
+		];
+
+		const hasDefaultProviderConfigKeys = (config: object) =>
+			Object.keys(config).find(key => defaultProviderConfigKeys.includes(key));
+
+		if (
+			hasDefaultProviderConfigKeys(storageConfig) &&
+			!storageConfig[DEFAULT_PROVIDER]
+		) {
+			storageConfig[DEFAULT_PROVIDER] = {};
+		}
+
+		Object.entries(storageConfig).forEach(([key, value]) => {
+			if (
+				key &&
+				defaultProviderConfigKeys.includes(key) &&
+				value !== undefined
+			) {
+				storageConfig[DEFAULT_PROVIDER][key] = value;
+				delete storageConfig[key];
+			}
+		});
+
+		// only update new values for each provider
+		Object.keys(storageConfig).forEach(providerName => {
+			if (typeof storageConfig[providerName] !== 'string') {
+				this._config[providerName] = {
+					...this._config[providerName],
+					...storageConfig[providerName],
+				};
+			}
+		});
+
+		this._pluggables.forEach(pluggable => {
+			pluggable.configure(this._config[pluggable.getProviderName()]);
+		});
+
+		if (this._pluggables.length === 0) {
+			this.addPluggable(new AWSS3Provider());
+		}
+
+		return this._config;
+	}
+
+	private getAbortController(): AbortController {
+		return new AbortController();
+	}
+
+	private updateRequestToBeCancellable(
+		request: Promise<any>,
+		abortController: AbortController
+	) {
+		this._abortControllerMap.set(request, abortController);
+	}
+
+	private isUploadTask(x: unknown): x is UploadTask {
+		return (
+			typeof x !== 'undefined' &&
+			typeof x['pause'] === 'function' &&
+			typeof x['resume'] === 'function'
+		);
 	}
 
 	/**
@@ -54,7 +218,19 @@ export class Storage extends InternalStorageClass {
 		request: Promise<any> | UploadTask,
 		message?: string
 	): void | Promise<boolean> {
-		return super.cancel(request, message);
+		if (request instanceof AWSS3UploadTask) {
+			return request._cancel();
+		}
+		const abortController = this._abortControllerMap.get(
+			request as Promise<any>
+		);
+		if (abortController) {
+			// TODO: [v6] clean up the aborted promise in the weak map.
+			// Not doing it yet to avoid breaking changes when users may abort a request twice.
+			abortController.abort(message);
+		} else {
+			logger.debug('The request does not map to any cancel token');
+		}
 	}
 
 	/**
@@ -75,7 +251,28 @@ export class Storage extends InternalStorageClass {
 		dest: Parameters<T['copy']>[1],
 		config?: StorageCopyConfig<T>
 	): StorageCopyOutput<T> {
-		return super.copy(src, dest, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			return Promise.reject(
+				'No plugin found in Storage for the provider'
+			) as StorageCopyOutput<T>;
+		}
+		const abortController = this.getAbortController();
+		if (typeof plugin.copy !== 'function') {
+			return Promise.reject(
+				`.copy is not implemented on provider ${plugin.getProviderName()}`
+			) as StorageCopyOutput<T>;
+		}
+		const responsePromise = plugin.copy(src, dest, {
+			...config,
+			abortSignal: abortController.signal,
+		});
+		this.updateRequestToBeCancellable(responsePromise, abortController);
+		return responsePromise as StorageCopyOutput<T>;
 	}
 
 	/**
@@ -93,14 +290,52 @@ export class Storage extends InternalStorageClass {
 	public get<
 		T extends StorageProvider | { [key: string]: any; download?: boolean }
 	>(key: string, config?: StorageGetConfig<T>): StorageGetOutput<T> {
-		return super.get(key, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			return Promise.reject(
+				'No plugin found in Storage for the provider'
+			) as StorageGetOutput<T>;
+		}
+		const abortController = this.getAbortController();
+		const responsePromise = plugin.get(key, {
+			...config,
+			abortSignal: abortController.signal,
+		});
+		this.updateRequestToBeCancellable(responsePromise, abortController);
+		return responsePromise as StorageGetOutput<T>;
+	}
+
+	public isCancelError(error: any) {
+		return isCancelError(error);
 	}
 
 	public getProperties<T extends StorageProvider | { [key: string]: any }>(
 		key: string,
 		config?: StorageGetPropertiesConfig<T>
 	): StorageGetPropertiesOutput<T> {
-		return super.getProperties(key, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			throw new Error('No plugin found with providerName');
+		}
+		const abortController = this.getAbortController();
+		if (typeof plugin.getProperties !== 'function') {
+			return Promise.reject(
+				`.getProperties is not implemented on provider ${plugin.getProviderName()}`
+			) as StorageGetPropertiesOutput<T>;
+		}
+		const responsePromise = plugin?.getProperties(key, {
+			...config,
+		});
+		this.updateRequestToBeCancellable(responsePromise, abortController);
+		return responsePromise as StorageGetPropertiesOutput<T>;
 	}
 	/**
 	 * Put a file in storage bucket specified to configure method
@@ -120,7 +355,25 @@ export class Storage extends InternalStorageClass {
 		object: Omit<PutObjectInput['Body'], 'ReadableStream' | 'Readable'>,
 		config?: StoragePutConfig<T>
 	): StoragePutOutput<T> {
-		return super.put(key, object, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			return Promise.reject(
+				'No plugin found in Storage for the provider'
+			) as StoragePutOutput<T>;
+		}
+		const abortController = this.getAbortController();
+		const response = plugin.put(key, object, {
+			...config,
+			abortSignal: abortController.signal,
+		});
+		if (!this.isUploadTask(response)) {
+			this.updateRequestToBeCancellable(response, abortController);
+		}
+		return response as StoragePutOutput<T>;
 	}
 
 	/**
@@ -137,7 +390,17 @@ export class Storage extends InternalStorageClass {
 		key: string,
 		config?: StorageRemoveConfig<T>
 	): StorageRemoveOutput<T> {
-		return super.remove(key, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			return Promise.reject(
+				'No plugin found in Storage for the provider'
+			) as StorageRemoveOutput<T>;
+		}
+		return plugin.remove(key, config) as StorageRemoveOutput<T>;
 	}
 
 	/**
@@ -154,7 +417,17 @@ export class Storage extends InternalStorageClass {
 		path: string,
 		config?: StorageListConfig<T>
 	): StorageListOutput<T> {
-		return super.list(path, config);
+		const provider = config?.provider || DEFAULT_PROVIDER;
+		const plugin = this._pluggables.find(
+			pluggable => pluggable.getProviderName() === provider
+		);
+		if (plugin === undefined) {
+			logger.debug('No plugin found with providerName', provider);
+			return Promise.reject(
+				'No plugin found in Storage for the provider'
+			) as StorageListOutput<T>;
+		}
+		return plugin.list(path, config) as StorageListOutput<T>;
 	}
 }
