--- packages/storage/src/internals/index.ts
@@ -1,3 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-export { InternalStorage } from './InternalStorage';
