+++ packages/storage/src/types/Provider.ts
@@ -20,36 +20,26 @@ export interface StorageProvider {
 	copy?(
 		src: StorageCopySource,
 		dest: StorageCopyDestination,
-		config?,
-		userAgentValue?: string
+		config?
 	): Promise<any>;
 
 	// configure your provider
 	configure(config: object): object;
 
 	// get object/pre-signed url from storage
-	get(key: string, options?, userAgentValue?: string): Promise<string | Object>;
+	get(key: string, options?): Promise<string | Object>;
 
 	// get properties of object
-	getProperties?(
-		key: string,
-		options?,
-		userAgentValue?: string
-	): Promise<Object>;
+	getProperties?(key: string, options?): Promise<Object>;
 
 	// upload storage object
-	put(
-		key: string,
-		object,
-		options?,
-		userAgentValue?: string
-	): Promise<Object> | UploadTask;
+	put(key: string, object, options?): Promise<Object> | UploadTask;
 
 	// remove object
-	remove(key: string, options?, userAgentValue?: string): Promise<any>;
+	remove(key: string, options?): Promise<any>;
 
 	// list objects for the path
-	list(path, options?, userAgentValue?: string): Promise<any>;
+	list(path, options?): Promise<any>;
 
 	// return 'Storage';
 	getCategory(): string;
@@ -70,17 +60,12 @@ export interface StorageProviderWithCopy extends StorageProvider {
 	copy(
 		src: StorageCopySource,
 		dest: StorageCopyDestination,
-		config?,
-		userAgentValue?: string
+		config?
 	): Promise<any>;
 }
 
 export interface StorageProviderWithGetProperties extends StorageProvider {
-	getProperties(
-		key: string,
-		options?,
-		userAgentValue?: string
-	): Promise<Object>;
+	getProperties(key: string, options?): Promise<Object>;
 }
 
 export type StorageProviderApi =
@@ -90,14 +75,3 @@ export type StorageProviderApi =
 	| 'remove'
 	| 'list'
 	| 'getProperties';
-
-// Map of api to index of options (config) parameter
-// Used to glean config type from StorageProvider
-export type StorageProviderApiOptionsIndexMap = {
-	copy: 2;
-	get: 1;
-	put: 2;
-	remove: 1;
-	list: 1;
-	getProperties: 1;
-};
